requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
apscheduler==3.11.0
asgiref==3.9.2
attrs==25.3.0
discord==2.3.2
discord.py==2.6.3
Django==5.2.6
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
PyNaCl==1.6.0
python-dotenv==1.1.1
regex==2025.9.18
sqlalchemy==2.0.44
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
yarl==1.20.1
